requirements.txt
fastapi
uvicorn
transformers
torch
langchain
sqlalchemy
psycopg2-binary
faiss-cpu
python-dotenv
pydantic
